+++ packages/components/src/Box.ts
@@ -1,0 +1,1 @@
+export { Box, BoxProps as ElementProps, StyleProps } from '@marigold/system';
