--- packages/components/src/Box/Box.stories.mdx
@@ -1,334 +1,0 @@
-import { ArgsTable, Canvas, Meta, Story } from '@storybook/addon-docs';
-import { Box } from './Box';
-import { Button } from '../Button';
-import { Label } from '../Label';
-
-<Meta
-  title="Components/Box"
-  argTypes={{
-    variant: {
-      control: {
-        type: 'text',
-      },
-      table: {
-        category: 'General',
-      },
-    },
-    display: {
-      control: {
-        type: 'text',
-      },
-      table: {
-        category: 'General',
-      },
-    },
-    height: {
-      control: {
-        type: 'text',
-      },
-      table: {
-        category: 'Width',
-      },
-    },
-    width: {
-      control: {
-        type: 'text',
-      },
-      table: {
-        category: 'Width',
-      },
-    },
-    minWidth: {
-      control: {
-        type: 'text',
-      },
-      table: {
-        category: 'Width',
-      },
-    },
-    maxWidth: {
-      control: {
-        type: 'text',
-      },
-      table: {
-        category: 'Width',
-      },
-    },
-    position: {
-      control: {
-        type: 'text',
-      },
-      table: {
-        category: 'Position',
-      },
-    },
-    top: {
-      control: {
-        type: 'text',
-      },
-      table: {
-        category: 'Position',
-      },
-    },
-    bottom: {
-      control: {
-        type: 'text',
-      },
-      table: {
-        category: 'Position',
-      },
-    },
-    right: {
-      control: {
-        type: 'text',
-      },
-      table: {
-        category: 'Position',
-      },
-    },
-    left: {
-      control: {
-        type: 'text',
-      },
-      table: {
-        category: 'Position',
-      },
-    },
-    zIndex: {
-      control: {
-        type: 'text',
-      },
-      table: {
-        category: 'Position',
-      },
-    },
-    p: {
-      control: {
-        type: 'text',
-      },
-      table: {
-        category: 'Padding',
-      },
-    },
-    px: {
-      control: {
-        type: 'text',
-      },
-      table: {
-        category: 'Padding',
-      },
-    },
-    py: {
-      control: {
-        type: 'text',
-      },
-      table: {
-        category: 'Padding',
-      },
-    },
-    pt: {
-      control: {
-        type: 'text',
-      },
-      table: {
-        category: 'Padding',
-      },
-    },
-    pb: {
-      control: {
-        type: 'text',
-      },
-      table: {
-        category: 'Padding',
-      },
-    },
-    pl: {
-      control: {
-        type: 'text',
-      },
-      table: {
-        category: 'Padding',
-      },
-    },
-    pr: {
-      control: {
-        type: 'text',
-      },
-      table: {
-        category: 'Padding',
-      },
-    },
-    m: {
-      control: {
-        type: 'text',
-      },
-      table: {
-        category: 'Margin',
-      },
-    },
-    mx: {
-      control: {
-        type: 'text',
-      },
-      table: {
-        category: 'Margin',
-      },
-    },
-    my: {
-      control: {
-        type: 'text',
-      },
-      table: {
-        category: 'Margin',
-      },
-    },
-    mt: {
-      control: {
-        type: 'text',
-      },
-      table: {
-        category: 'Margin',
-      },
-    },
-    mb: {
-      control: {
-        type: 'text',
-      },
-      table: {
-        category: 'Margin',
-      },
-    },
-    ml: {
-      control: {
-        type: 'text',
-      },
-      table: {
-        category: 'Margin',
-      },
-    },
-    mr: {
-      control: {
-        type: 'text',
-      },
-      table: {
-        category: 'Margin',
-      },
-    },
-    flexDirection: {
-      control: {
-        type: 'text',
-      },
-      table: {
-        category: 'Flex',
-      },
-    },
-    flexWrap: {
-      control: {
-        type: 'text',
-      },
-      table: {
-        category: 'Flex',
-      },
-    },
-    flexShrink: {
-      control: {
-        type: 'text',
-      },
-      table: {
-        category: 'Flex',
-      },
-    },
-    flexGrow: {
-      control: {
-        type: 'text',
-      },
-      table: {
-        category: 'Flex',
-      },
-    },
-    alignItems: {
-      control: {
-        type: 'text',
-      },
-      table: {
-        category: 'Flex',
-      },
-    },
-    justifyContent: {
-      control: {
-        type: 'text',
-      },
-      table: {
-        category: 'Flex',
-      },
-    },
-    bg: {
-      control: {
-        type: 'text',
-      },
-      table: {
-        category: 'Other',
-      },
-    },
-    border: {
-      control: {
-        type: 'text',
-      },
-      table: {
-        category: 'Other',
-      },
-    },
-    borderRadius: {
-      control: {
-        type: 'text',
-      },
-      table: {
-        category: 'Other',
-      },
-    },
-    boxShadow: {
-      control: {
-        type: 'text',
-      },
-      table: {
-        category: 'Other',
-      },
-    },
-    opacity: {
-      control: {
-        type: 'text',
-      },
-      table: {
-        category: 'Other',
-      },
-    },
-    overflow: {
-      control: {
-        type: 'text',
-      },
-      table: {
-        category: 'Other',
-      },
-    },
-    transition: {
-      control: {
-        type: 'text',
-      },
-      table: {
-        category: 'Other',
-      },
-    },
-  }}
-/>
-
-# Box-Component
-
-## For custom components
-
-export const Template = args => <Box {...args}>Style me!</Box>;
-
-<Canvas>
-  <Story name="Default">{Template.bind({})}</Story>
-</Canvas>
-
-<ArgsTable story="Default" />
