--- packages/components/src/Box/index.ts
@@ -1,1 +1,0 @@
-export * from './Box';
