+++ packages/system/src/index.ts
@@ -1,4 +1,4 @@
-export * from './Element';
+export * from './Box';
 export * from './Global';
 export * from './types';
 export * from './useTheme';
